加入网关及权限框架
