+++ aggregate-gateway/src/main/java/com/xin/gateway/config/CosConfig.java
@@ -1,0 +1,29 @@
+package com.xin.gateway.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.cors.CorsConfiguration;
+import org.springframework.web.cors.reactive.CorsWebFilter;
+import org.springframework.web.cors.reactive.UrlBasedCorsConfigurationSource;
+import org.springframework.web.util.pattern.PathPatternParser;
+
+/**
+ * @author xin
+ * @create 2021/12/06 0:16
+ * @description
+ **/
+@Configuration
+public class CosConfig {
+
+    @Bean
+    public CorsWebFilter corsWebFilter(){
+        CorsConfiguration config = new CorsConfiguration();
+        config.setAllowCredentials(true); //是否允许携带cookie
+        config.addAllowedOrigin("*"); //可接受的域，是一个具体域名或者*（代表任意域名）
+        config.addAllowedHeader("*"); //允许携带的头
+        config.addAllowedMethod("*"); //允许访问的方式
+        UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource(new PathPatternParser());
+        source.registerCorsConfiguration("/**",config);
+        return new CorsWebFilter(source);
+    }
+}
