+++ aggregate-gateway/src/main/resources/bootstrap.yml
@@ -1,11 +1,11 @@
 spring:
   profiles:
     active: dev
+  application:
+    name: aggregate-gateway
   cloud:
     nacos:
       config:
         server-addr: 101.43.61.239:8848
         namespace: 690e45ec-c5de-4587-bcb1-f0aad9e1d670
         file-extension: yaml
-  application:
-    name: aggregate-info
