--- aggregate-info/src/main/java/com/info/aggregateinfo/constant/Constants.java
@@ -1,11 +1,0 @@
-package com.info.aggregateinfo.constant;
-
-/**
- * @author xin
- * @create 2021/11/18 1:59
- * @description
- **/
-public interface Constants {
-
-    String YZM = "验证码";
-}
