--- aggregate-info/src/main/java/com/info/aggregateinfo/mapper/MailInfoMapper.java
@@ -1,16 +1,0 @@
-package com.info.aggregateinfo.mapper;
-
-import com.baomidou.mybatisplus.core.mapper.BaseMapper;
-import com.info.aggregateinfo.pojo.entity.MailInfo;
-
-/**
- * <p>
- * 邮件信息 Mapper 接口
- * </p>
- *
- * @author xin
- * @since 2021-11-21
- */
-public interface MailInfoMapper extends BaseMapper<MailInfo> {
-
-}
