--- aggregate-info/src/main/java/com/info/aggregateinfo/pojo/dto/AggregateAdminDTO.java
@@ -1,46 +1,0 @@
-package com.info.aggregateinfo.pojo.dto;
-
-import com.baomidou.mybatisplus.annotation.TableField;
-import io.swagger.annotations.ApiModel;
-import io.swagger.annotations.ApiModelProperty;
-import lombok.Data;
-
-import java.time.LocalDateTime;
-
-/**
- * @author xin
- * @create 2021/11/17 1:47
- * @description
- **/
-@Data
-@ApiModel(value = "用户注册信息")
-public class AggregateAdminDTO {
-
-    @ApiModelProperty("用户名")
-    private String username;
-
-    @ApiModelProperty("密码")
-    private String password;
-
-    @ApiModelProperty("头像")
-    private String icon;
-
-    @ApiModelProperty("邮箱")
-    private String email;
-
-    @TableField("nick_name")
-    @ApiModelProperty("昵称")
-    private String nickName;
-
-    @ApiModelProperty("备注信息")
-    private String note;
-
-    @ApiModelProperty("创建时间")
-    @TableField("create_time")
-    private LocalDateTime createTime;
-
-    @ApiModelProperty("最后登录时间")
-    @TableField("login_time")
-    private LocalDateTime loginTime;
-
-}
