+++ aggregate-info/src/main/java/com/xin/aggregateInfo/config/WebSecurityConfig.java
@@ -1,0 +1,49 @@
+package com.xin.aggregateInfo.config;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
+
+/**
+ * @author xin
+ * @create 2021/11/24 0:06
+ * @description
+ **/
+@Configuration
+public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
+
+    @Autowired
+    private UserDetailsService userDetailsService;
+
+    @Override
+    protected void configure(AuthenticationManagerBuilder auth) throws Exception {
+        auth.userDetailsService(userDetailsService).passwordEncoder(passwordEncoder());
+    }
+
+    @Bean
+    PasswordEncoder passwordEncoder(){
+        return new BCryptPasswordEncoder();
+    }
+
+    @Override
+    protected void configure(HttpSecurity http) throws Exception {
+        // 自定义登录页面
+        http.formLogin()
+                // 登录页面设置
+                .loginPage("http://localhost:3000/login")
+                .loginProcessingUrl("/admin/login")
+                .defaultSuccessUrl("http://localhost:3000")
+                .and().authorizeRequests()
+                // 设置那些路径可以直接访问，不需要认证
+                .antMatchers("http://localhost:3000/login").permitAll()
+                .anyRequest().authenticated()
+                // 关闭csrf防护
+                .and().csrf().disable();
+    }
+}
