+++ aggregate-info/src/main/java/com/xin/aggregateInfo/constant/Constants.java
@@ -1,0 +1,11 @@
+package com.xin.aggregateInfo.constant;
+
+/**
+ * @author xin
+ * @create 2021/11/18 1:59
+ * @description
+ **/
+public interface Constants {
+
+    String YZM = "验证码";
+}
