+++ aggregate-info/src/main/java/com/xin/aggregateInfo/mapper/AggregateAdminMapper.java
@@ -1,0 +1,18 @@
+package com.xin.aggregateInfo.mapper;
+
+import com.xin.aggregateInfo.pojo.entity.AggregateAdmin;
+import com.baomidou.mybatisplus.core.mapper.BaseMapper;
+import org.springframework.stereotype.Repository;
+
+/**
+ * <p>
+ * 用户信息表	 Mapper 接口
+ * </p>
+ *
+ * @author xin
+ * @since 2021-11-14
+ */
+@Repository
+public interface AggregateAdminMapper extends BaseMapper<AggregateAdmin> {
+
+}
