+++ aggregate-info/src/main/java/com/xin/aggregateInfo/mapper/MailInfoMapper.java
@@ -1,0 +1,16 @@
+package com.xin.aggregateInfo.mapper;
+
+import com.baomidou.mybatisplus.core.mapper.BaseMapper;
+import com.xin.aggregateInfo.pojo.entity.MailInfo;
+
+/**
+ * <p>
+ * 邮件信息 Mapper 接口
+ * </p>
+ *
+ * @author xin
+ * @since 2021-11-21
+ */
+public interface MailInfoMapper extends BaseMapper<MailInfo> {
+
+}
