--- aggregate-info/src/test/java/com/info/aggregateinfo/AggregateInfoApplicationTests.java
@@ -1,13 +1,0 @@
-package com.info.aggregateinfo;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class AggregateInfoApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
