+++ aggregate-info/src/test/java/com/xin/aggregateInfo/AggregateInfoApplicationTests.java
@@ -1,0 +1,13 @@
+package com.xin.aggregateInfo;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class AggregateInfoApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
