+++ base-common/src/main/java/com/xin/constant/Constants.java
@@ -1,0 +1,11 @@
+package com.xin.constant;
+
+/**
+ * @author xin
+ * @create 2021/11/18 1:59
+ * @description
+ **/
+public interface Constants {
+
+    int STATUS = 1;
+}
