--- base-common/src/main/java/constant/Constants.java
@@ -1,11 +1,0 @@
-package constant;
-
-/**
- * @author xin
- * @create 2021/11/18 1:59
- * @description
- **/
-public interface Constants {
-
-    int STATUS = 1;
-}
